+++ src/main/kotlin/org/deniskusakin/aem/groovyconsoleplugin/services/http/model/GroovyConsoleOutput.kt
@@ -1,0 +1,3 @@
+package org.deniskusakin.aem.groovyconsoleplugin.services.http.model
+
+data class GroovyConsoleOutput(val output: String, val runningTime: String, val exceptionStackTrace: String)