+++ src/main/kotlin/org/deniskusakin/aem/groovyconsoleplugin/services/http/model/GroovyConsoleResponseHandler.kt
@@ -1,0 +1,7 @@
+package org.deniskusakin.aem.groovyconsoleplugin.services.http.model
+
+interface GroovyConsoleResponseHandler {
+    fun onSuccess(output: GroovyConsoleOutput)
+
+    fun onFail(th: Throwable)
+}