+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 # aem-groovyconsole-plugin Changelog
 ## [Unreleased]
 ## [0.3.0]
+### Added
  - Added support for methods and bindings in groovy editor
+### Changed
  - Reworked Configuration UI
- - Updated storage for username/password
+ - Updated storage for username/password. Need to reconfigure servers.
 ## [0.2.11]
 ### Changed
 - Fix compatibility issue. Plugin is now available for IDEA 2022.2